Refactor: Add Graph class
* 기존의 CandleChart 클래스에서 그래프 부분을 별도의 클래스로 분리
* 가격축, 시간축 캔버스와 같이 관리하기 위함

-
