+++ packages/dogyeong/src/frontend/chart/index.ts
@@ -2,7 +2,5 @@ import CandleChart from '@/chart/CandleChart';
 import { createCanvas } from '@/chart/utils';
 
 export const createChart = ($container: HTMLElement) => {
-  const canvas = createCanvas($container);
-
-  return new CandleChart(canvas);
+  return new CandleChart($container);
 };
